update wrbFormAppSelector data fetching component
